+++ CMakeLists.txt
@@ -9,4 +9,5 @@ aux_source_directory(${PROJECT_SOURCE_DIR}/glad/src DIR_GLAD_SRCS)
 add_executable(${PROJECT_NAME} ${DIR_MAIN_SRCS} ${DIR_GLAD_SRCS})
 target_include_directories(${PROJECT_NAME} PRIVATE ${PROJECT_SOURCE_DIR}/glad/include)
 target_include_directories(${PROJECT_NAME} PRIVATE ${PROJECT_SOURCE_DIR}/glfw/include)
+target_include_directories(${PROJECT_NAME} PRIVATE ${PROJECT_SOURCE_DIR}/src)
 target_link_libraries(${PROJECT_NAME} PUBLIC glfw)
